Moved eval scripts
